requirements.txt
asgiref==3.8.1
Django~=5.1.5
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
PyYAML~=6.0.2